data/ml/requirements.model.txt
absl-py==1.4.0
boto3==1.26.135
fsspec==2023.5.0
gin-config==0.5.0
lightning==2.2.0
lion-pytorch==0.1.2
loguru==0.7.0
numpy==1.24.3
pandas==2.0.0
pyarrow==12.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rich==13.3.5
sagemaker==2.197.0
s3fs==2023.5.0
scipy==1.10.1
tensorboard==2.13.0
tensorflow-io==0.32.0; sys_platform != 'darwin' or platform_machine != 'arm64'
torchsampler==0.1.2
transformers==4.28.1
